cloud_run/requirements.txt
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4
google-cloud-logging==3.5.0
pandas==2.0.3
pytz==2023.3
requests==2.31.0
retry==0.9.2
flask==2.3.3
gunicorn==21.2.0 